CORE-237 moved confirmation dialog to libs
