+++ frontend/angular.json
@@ -1123,6 +1123,41 @@
           "style": "scss"
         }
       }
+    },
+    "shared-ui-confirmation-dialog": {
+      "projectType": "library",
+      "root": "libs/shared/ui-confirmation-dialog",
+      "sourceRoot": "libs/shared/ui-confirmation-dialog/src",
+      "prefix": "qro",
+      "architect": {
+        "lint": {
+          "builder": "@angular-devkit/build-angular:tslint",
+          "options": {
+            "tsConfig": [
+              "libs/shared/ui-confirmation-dialog/tsconfig.lib.json",
+              "libs/shared/ui-confirmation-dialog/tsconfig.spec.json"
+            ],
+            "exclude": [
+              "**/node_modules/**",
+              "!libs/shared/ui-confirmation-dialog/**"
+            ]
+          }
+        },
+        "test": {
+          "builder": "@nrwl/jest:jest",
+          "options": {
+            "jestConfig": "libs/shared/ui-confirmation-dialog/jest.config.js",
+            "tsConfig": "libs/shared/ui-confirmation-dialog/tsconfig.spec.json",
+            "passWithNoTests": true,
+            "setupFile": "libs/shared/ui-confirmation-dialog/src/test-setup.ts"
+          }
+        }
+      },
+      "schematics": {
+        "@nrwl/angular:component": {
+          "style": "scss"
+        }
+      }
     }
   },
   "cli": {
