+++ frontend/libs/shared/ui-confirmation-dialog/README.md
@@ -1,0 +1,7 @@
+# shared-ui-confirmation-dialog
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-ui-confirmation-dialog` to execute the unit tests.
