+++ frontend/libs/shared/ui-confirmation-dialog/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/shared-ui-confirmation-dialog.module';
+export * from './lib/confirmation-dialog/confirmation-dialog.component';
