+++ frontend/libs/shared/ui-confirmation-dialog/src/lib/confirmation-dialog/confirmation-dialog.component.scss
