+++ frontend/libs/shared/ui-confirmation-dialog/src/lib/shared-ui-confirmation-dialog.module.spec.ts
@@ -1,0 +1,14 @@
+import { async, TestBed } from '@angular/core/testing';
+import { SharedUiConfirmationDialogModule } from './shared-ui-confirmation-dialog.module';
+
+describe('SharedUiConfirmationDialogModule', () => {
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [SharedUiConfirmationDialogModule],
+    }).compileComponents();
+  }));
+
+  it('should create', () => {
+    expect(SharedUiConfirmationDialogModule).toBeDefined();
+  });
+});
