+++ frontend/libs/shared/ui-confirmation-dialog/src/lib/shared-ui-confirmation-dialog.module.ts
@@ -1,15 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ConfirmationDialogComponent } from './confirmation-dialog.component';
 import { SharedUiMaterialModule } from '@qro/shared/ui-material';
+import { ConfirmationDialogComponent } from './confirmation-dialog/confirmation-dialog.component';
 
 @NgModule({
-  imports: [
-    CommonModule,
-    SharedUiMaterialModule
-  ],
+  imports: [CommonModule, SharedUiMaterialModule],
   declarations: [ConfirmationDialogComponent],
   exports: [ConfirmationDialogComponent],
   entryComponents: [ConfirmationDialogComponent]
 })
-export class ConfirmationDialogModule { }
+export class SharedUiConfirmationDialogModule { }
