+++ frontend/libs/shared/ui-confirmation-dialog/tslint.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../../tslint.json",
+  "rules": {
+    "directive-selector": [true, "attribute", "qro", "camelCase"],
+    "component-selector": [true, "element", "qro", "kebab-case"]
+  },
+  "linterOptions": {
+    "exclude": ["!**/*"]
+  }
+}
