+++ frontend/libs/shared/util/src/lib/prevent-unsaved-changes.guard.ts
@@ -3,7 +3,7 @@ import { CanDeactivate } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { ConfirmationDialogComponent } from '../../../../../apps/quarano-frontend/src/app/ui/confirmation-dialog/confirmation-dialog.component';
+import { ConfirmationDialogComponent } from '../../../ui-confirmation-dialog/src/lib/confirmation-dialog/confirmation-dialog.component';
 
 export interface DeactivatableComponent {
   canDeactivate: () => boolean | Observable<boolean>;
