+++ frontend/tsconfig.json
@@ -73,7 +73,10 @@
         "libs/administration/accounts/shell/src/index.ts"
       ],
       "@qro/auth/api": ["libs/auth/api/src/index.ts"],
-      "@qro/shared/ui-alert": ["libs/shared/ui-alert/src/index.ts"]
+      "@qro/shared/ui-alert": ["libs/shared/ui-alert/src/index.ts"],
+      "@qro/shared/ui-confirmation-dialog": [
+        "libs/shared/ui-confirmation-dialog/src/index.ts"
+      ]
     }
   },
   "exclude": ["node_modules", "tmp"]
